Updated project 1 file for movie class
Still need to add the showing method and coming method but the user can now add all the information needed to add a movie to a list. Currently the method will add all the information about that movie being added to an array. Plus there is error checking for invalid date inputs and if the release date is earlier than received.